Fix #5 and #7
Fixes the issue where `next` doesn't always keep the same search query. This is solved by always passing to the search the current value of the textbox

Also keeping track of the laste request via an int value now

Plus we don't do a query if there is nothing to search for. This was causing an issue where the user could press down when there was nothing in the search box and shit was all fall over